backend/requirements.txt
fastapi
uvicorn
python-multipart
websockets
torch
torchvision
pillow
numpy
tqdm
requests
